moves suite2p_registration to modules
