+++ tests/modules/suite2p_registration/test_suite2p_registration.py
@@ -14,7 +14,7 @@ sys.modules['suite2p'] = Mock()
 sys.modules['suite2p.registration.rigid'] = Mock()
 from ophys_etl.modules.suite2p_wrapper.schemas import \
         Suite2PWrapperSchema, Suite2PWrapperOutputSchema  # noqa: E402
-import ophys_etl.pipelines.suite2p_registration as s2preg  # noqa
+import ophys_etl.modules.suite2p_registration.__main__ as s2preg  # noqa
 
 
 class MockSuite2PWrapper(argschema.ArgSchemaParser):
@@ -51,41 +51,6 @@ class MockSuite2PWrapper(argschema.ArgSchemaParser):
 
 
 @pytest.mark.suite2p_only
-@pytest.mark.parametrize(
-        ("excess_indices", "deltas", "thresh", "expected_indices"), [
-            ([], [], 10, []),
-            ([123, 567], [20, 20], 10, [123, 567]),
-            ([123, 567], [8, 20], 10, [567]),
-            ([123, 567], [-20, -20], 10, [123, 567]),
-            ([123, 567], [-8, -20], 10, [567]),
-            ([123, 567, 1234, 5678], [-20, -8, 8, 20], 10, [123, 5678])
-            ])
-def test_identify_and_clip_outliers(excess_indices, deltas,
-                                    thresh, expected_indices):
-    frame_index = np.arange(10000)
-    # long-range drifts
-    baseline = 20.0 * np.cos(2.0 * np.pi * frame_index / 500)
-    additional = np.zeros_like(baseline)
-    for index, delta in zip(excess_indices, deltas):
-        additional[index] += delta
-
-    data, indices = s2preg.identify_and_clip_outliers(
-            baseline + additional,
-            10,
-            thresh)
-
-    # check that the outliers were identified
-    assert set(indices) == set(expected_indices)
-    # check that the outlier values were clipped to within
-    # the threshold of the underlying trend data
-    # with a small delta value because the median-filtered data
-    # is not quite ever the same as the baseline data
-    deltas = np.abs(data[expected_indices] - baseline[expected_indices])
-    small_delta = 1.0
-    assert np.all(deltas < thresh + small_delta)
-
-
-@pytest.mark.suite2p_only
 @pytest.mark.parametrize("mock_ops_data", [
     {"Lx": 0, "Ly": 0, "nframes": 5, "xrange": 0, "yrange": 0,
      "xoff": [1, 2, 3, 4, 5], "yoff": [5, 4, 3, 2, 1],
@@ -110,8 +75,9 @@ def test_suite2p_registration(tmp_path, mock_ops_data):
             "output_json": str(outj_path)}
 
     with patch.object(MockSuite2PWrapper, "mock_ops_data", mock_ops_data):
-        with patch('ophys_etl.pipelines.suite2p_registration.Suite2PWrapper',
-                   MockSuite2PWrapper):
+        with patch(
+            'ophys_etl.modules.suite2p_registration.__main__.Suite2PWrapper',
+                MockSuite2PWrapper):
             reg = s2preg.Suite2PRegistration(input_data=args, args=[])
             reg.run()
